+++ internal/modules/warehouse/warehouse.dto.go
@@ -33,7 +33,7 @@ type WarehouseTableByYearResponse struct {
 	TotalWarehouses       int                  `json:"total_warehouses"`
 	TotalActiveWarehouses int                  `json:"total_active_warehouses"`
 	TotalStocksPods       int                  `json:"total_stocks_pods"`
-	TotalStocksWeights    int                  `json:"total_stocks_weights"`
+	TotalStocksWeights    float64              `json:"total_stocks_weights"`
 	WarehouseTable        []WarehouseTableItem `json:"warehouse_table"`
 }
 
@@ -42,12 +42,12 @@ type WarehouseTableItem struct {
 	WarehouseName string `json:"warehouse_name"`
 	ActiveStatus  bool   `json:"active_status"`
 
-	TotalPods    int `json:"total_pods"`
-	TotalWeights int `json:"total_weights"`
+	TotalPods    int     `json:"total_pods"`
+	TotalWeights float64 `json:"total_weights"`
 
-	DistributedPods    int `json:"distributed_pods"`
-	DistributedWeights int `json:"distributed_weights"`
+	DistributedPods    int     `json:"distributed_pods"`
+	DistributedWeights float64 `json:"distributed_weights"`
 
-	RemainingPods    int `json:"remaining_pods"`
-	RemainingWeights int `json:"remaining_weights"`
+	RemainingPods    int     `json:"remaining_pods"`
+	RemainingWeights float64 `json:"remaining_weights"`
 }
